Upload the congratulation remark
Original congratulation remark from Dr Schumie and Dr Madigan

The Korean Chapter of OHDSI makes many valuable contributions to OHDSI in many different areas, ranging from extensions of the Common Data Model (CDM), and development of novel methods, to applications of the OHDSI tools to answering important clinical questions. The CDM has been widely adopted throughout Korea, forming a vibrant distributed research network. Not only does the Korean Chapter organize OHDSI Symposia in Korea, welcoming all from across the world, but we also always find many of our Korean friends in other OHDSI Symposia across the globe. The community spirit that is at the heart of OHDSI is very much alive in Korea!

The Book of OHDSI was written by the OHDSI community, including several members from Korea, for the OHDSI community. We congratulate the Korean Chapter on their translation of the Book to Korean. We are convinced this Korean version will be of great value to the Korean OHDSI community.